requirements.txt
cachetools
pillow
psutil
asyncpg

git+https://github.com/Gorialis/jishaku
git+https://github.com/Rapptz/discord.py